indent-bars-ts: use the new jit/font-lock approach for scope bars
indent-bars now handles all of the bar searching and drawing itself,
and we simply configure it to i) use the appropriate in/out-of-scope
style, ii) skip font-lock when it's not needed.  For the latter we
advise `jit-lock-context' and provide a
`jit-lock-after-change-extend-region-function', both of which simply
mark text with a new `indent-bars-font-lock-pending' property.

We no longer need range intersection capability.