Improve depth=0 behavior, relevant for starting-column=0
If starting-column=nil, no column is drawn at column 0, so there is no
distinction between non-blank lines lacking any initial white-space
for which a bar should or should not be highlighted.  Basically "no
highlighted bar" and "there is no bar to highlight" were not
discriminated.  But if starting-column=0, these are distinct states.

This updates that logic to respect depth=0 as indicating "no
highlight", which is the initial state.  Bug #100.